pyproject.toml
[tool.poetry]
name = "jwt"
version = "0.1.0"
description = ""
authors = ["phuyalbish <phuyalbish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
mysql-connector-python = "^8.4.0"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
